проброс jwt между сервисами

+++ services/rental/src/logic/entity/class.ts
@@ -1,25 +1,52 @@
-import { EntityLogic, EntityPaginationData, EntityPaginationFilter, EntityStorage } from '@rsoi-lab2/library';
+import { ConfigurableLogic, EntityLogic, EntityPaginationData, EntityPaginationFilter, EntityStorage, LogicOptions } from '@rsoi-lab2/library';
 import { Rental, RentalFilter, RentalId } from './interface';
 
-export class RentalLogic implements EntityLogic<Rental, RentalFilter, RentalId> {
-    constructor(storage: EntityStorage<Rental, RentalFilter, RentalId>) {
+export class RentalLogic implements 
+    EntityLogic<Rental, RentalFilter, RentalId>,
+    ConfigurableLogic<RentalLogic> 
+{
+    constructor(
+        storage: EntityStorage<Rental, RentalFilter & {username: string}, RentalId>,
+        options?: LogicOptions
+    ) {
         this.storage = storage;
+        this.options = options;
     }
 
     getIdType(): 'string' | 'number' {
         return this.storage.getIdType();
     }
 
+    withOptions(options: LogicOptions): ConfigurableLogic<RentalLogic, LogicOptions> {
+        return new RentalLogic(this.storage, options);
+    }
+
     async getOne(id: RentalId): Promise<Required<Rental> | null> {
+        if (this.options == null) {
+            return Promise.reject(new Error('Rental logic not configured'));
+        }
+
         this.validateId(id);
 
-        return this.storage.getOne(id);
+        const rental = await this.storage.getOne(id);
+
+        if (rental == null || rental.username !== this.options.username) {
+            return null;
+        }
+
+        return rental;
     }
 
     async getMany(filter: RentalFilter): Promise<Array<Required<Rental>>> {
+        if (this.options == null) {
+            return Promise.reject(new Error('Rental logic not configured'));
+        }
+
         this.validateFilter(filter);
 
-        return this.storage.getMany(filter);
+        const fullFilter: RentalFilter = {...filter, username: this.options.username};
+
+        return this.storage.getMany(fullFilter);
     }
 
     supportsPagination(): boolean {
@@ -27,26 +54,56 @@ export class RentalLogic implements EntityLogic<Rental, RentalFilter, RentalId> 
     }
 
     async getPaginatedMany(filter: RentalFilter & EntityPaginationFilter): Promise<EntityPaginationData<Required<Rental>>> {
+        if (this.options == null) {
+            return Promise.reject(new Error('Rental logic not configured'));
+        }
+
         this.validateFilter(filter);
 
-        return this.storage.getPaginatedMany(filter);
+        const fullFilter: RentalFilter & EntityPaginationFilter & {username: string} = {...filter, username: this.options.username};
+
+        return this.storage.getPaginatedMany(fullFilter);
     }
 
     async create(entity: Rental): Promise<Required<Rental>> {
+        if (this.options == null) {
+            return Promise.reject(new Error('Rental logic not configured'));
+        }
+
         this.validateEntity(entity);
 
-        return this.storage.create(entity);
+        return this.storage.create({...entity, username: this.options.username});
     }
 
     async update(id: RentalId, update: Partial<Rental>): Promise<Required<Rental>> {
+        if (this.options == null) {
+            return Promise.reject(new Error('Rental logic not configured'));
+        }
+
         this.validateId(id);
         this.validatePartialEntity(update);
+
+        const rental = await this.storage.getOne(id);
+
+        if (rental == null) {
+            return Promise.reject(new Error('Did not update'));
+        }
 
         return this.storage.update(id, update);
     }
 
     async delete(id: RentalId): Promise<boolean> {
+        if (this.options == null) {
+            return Promise.reject(new Error('Rental logic not configured'));
+        }
+
         this.validateId(id);
+
+        const rental = await this.storage.getOne(id);
+
+        if (rental == null) {
+            return false;
+        }
 
         return this.storage.delete(id);
     }
@@ -58,7 +115,7 @@ export class RentalLogic implements EntityLogic<Rental, RentalFilter, RentalId> 
     }
 
     validateEntity(value: Rental): void {
-        for(const key of ['carUid', 'paymentUid', 'username', 'dateFrom', 'dateTo']) {
+        for(const key of ['carUid', 'paymentUid', 'dateFrom', 'dateTo']) {
             if (!value.hasOwnProperty(key)) {
                 throw new Error(`Invalid rental: must contain value in ${key} key`);
             }
@@ -67,10 +124,8 @@ export class RentalLogic implements EntityLogic<Rental, RentalFilter, RentalId> 
         this.validatePartialEntity(value);
     }
 
-    validateFilter(value: RentalFilter): void {
-        if(value.username.length === 0) {
-            throw new Error('Invalid rental filter: has empty username');
-        }
+    validateFilter(_: RentalFilter): void {
+        
     }
 
     validatePartialEntity(value: Partial<Rental>): void {
@@ -82,12 +137,17 @@ export class RentalLogic implements EntityLogic<Rental, RentalFilter, RentalId> 
             throw new Error('Invalid rental: has wrong dates order');
         }
 
-        for(const key of ['carUid', 'paymentUid', 'username']) {
+        for(const key of ['carUid', 'paymentUid']) {
             if (value.hasOwnProperty(key) && value[key].length === 0) {
                 throw new Error(`Invalid rental: must contain non-empty string in key ${key}`);
             }
         }
+
+        if (value.hasOwnProperty('username')) {
+            throw new Error('Invalid rental: has username');
+        }
     }
 
     private storage: EntityStorage<Rental, RentalFilter, RentalId>;
+    private options?: LogicOptions;
 }
